docker_api/brief_FastAPI/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
bcrypt==4.0.1
catboost==1.2.7
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
executing==2.2.0
Faker==36.1.1
fastapi==0.115.8
fastapi-cli==0.0.7
fonttools==4.56.0
graphviz==0.20.3
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.27.1
nest-asyncio==1.6.0
numpy==1.26.4
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
scipy==1.15.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlmodel==0.0.22
stack-data==0.6.3
starlette==0.45.3
threadpoolctl==3.5.0
tornado==6.4.2
traitlets==5.14.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0
pyodbc
faker